Turn cache-related functions into `FSCache` methods
